move lenses code to commonMain + add test
